Moved startfiles to own folder and deleted irrelevant file
