+++ navigation/RootStackNavigator.tsx
@@ -4,12 +4,12 @@ import React from 'react';
 import { useColorScheme } from 'react-native';
 import ChoreDetailScreen from '../screens/ChoreDetailScreen';
 import DistributeChoreScreen from '../screens/DistributeChoreScreen';
-import LoginScreen from '../screens/LoginScreen';
+import LoginScreen from '../screens/start/LoginScreen';
 import MemberDetailScreen from '../screens/MemberDetailScreen';
 import MemberScreen from '../screens/MemberScreen';
 import NoHouseholdScreen from '../screens/NoHouseholdScreen';
-import RegisterScreen from '../screens/RegisterScreen';
-import StartScreen from '../screens/StartScreen';
+import RegisterScreen from '../screens/start/RegisterScreen';
+import StartScreen from '../screens/start/StartScreen';
 import ChoresTabNavigator from './ChoresTabNavigator';
 import HouseholdChoresTabNavigator from './HouseholdChoresTabNavigator';
 import PieChartTabNavigator from './PieChartTabNavigator';
