--- screens/SplashScreen.tsx
@@ -1,8 +1,0 @@
-import { Text } from 'react-native';
-import React from 'react';
-
-export default function SplashScreen() {
-    return (
-        <Text>Splash screen</Text>
-    );
-}
