+++ screens/start/RegisterScreen.tsx
@@ -2,9 +2,9 @@ import { useTheme } from 'react-native-paper';
 import React, { useState } from 'react';
 import { StyleSheet, Text, TextInput, View } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { CustomNavigateButton } from '../component/CustomNavigateButton';
-import { RootStackScreenProps } from '../navigation/RootStackNavigator';
-import { useAppDispatch } from '../store/store';
+import { CustomNavigateButton } from '../../component/CustomNavigateButton';
+import { RootStackScreenProps } from '../../navigation/RootStackNavigator';
+import { useAppDispatch } from '../../store/store';
 import uuid from 'react-native-uuid';
 
 export default function RegisterScreen({ navigation }: RootStackScreenProps<'Register'>) {
